Stop filter from re-rendering on every keystroke
